Add comments to all classes & cleanup
Add comments. Delete classes and some of the methods that are no
longer being utilised. Change 'About' and 'Help' content.
Configure log4j to output info to the console and a log file. Remove
all occurrences of System.out.println() in the code. Add Logger.INFO()
to some methods instead.